Fix geolocation accuracy parameter type. Add DevTools section to README.md. Update LICENSE year. Add Ukrainian localization and remove redundant value from Russian localization

+++ src/main/java/aquality/selenium/browser/devtools/EmulationHandling.java
@@ -44,7 +44,7 @@ public class EmulationHandling {
      * @param longitude Longitude of location
      * @param accuracy Accuracy of the location
      */
-    public void setGeolocationOverride(double latitude, double longitude, int accuracy) {
+    public void setGeolocationOverride(double latitude, double longitude, double accuracy) {
         setGeolocationOverride(Optional.of(latitude), Optional.of(longitude), Optional.of(accuracy));
     }
